requirements.txt
couchbase>=4.3.0
click>=8.1.0
pyyaml>=6.0
rich>=13.0.0
